creation and cancelled order

+++ models.py
@@ -35,7 +35,7 @@ class  Order(Base):
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     status = Column(String, nullable=False)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    total = Column(Float, nullable=False)
+    total = Column(Integer, nullable=False)
     #ITEMS ORDERED RELATIONSHIP
 
     def __init__(self, status="PENDING", user_id: int = None, total=0):
